% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCR-package.R
\docType{package}
\name{HCR-package}
\alias{HCR}
\alias{HCR-package}
\title{HCR: Calculate Hydrologic Connectivity to Rivers}
\description{
Uses least cost routing to identify the most connected watershed pour point to a point in the coastal zone based on raster representing a river plume.
}
\author{
\strong{Maintainer}: Kelly Hondula \email{khondula@asu.edu}

}
\keyword{internal}
