% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_utils.R
\name{make_raster_reachable}
\alias{make_raster_reachable}
\title{Make raster reachable}
\usage{
make_raster_reachable(rast_to_fix, my_src_pt, src_pt_index = 1)
}
\arguments{
\item{rast_to_fix}{A raster with isolated water pixels}

\item{my_src_pt}{Point vector object for ocean point of interest}

\item{src_pt_index}{Row of ocean point of interest in my_src_pt, default = 1}
}
\value{
Raster object with pixels removed that arent connected to the source point
}
\description{
Make raster reachable
}
\examples{
# reachable_raster <- make_raster_reachable(rat)
}
