% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_utils.R
\name{sample_lines_and_extract}
\alias{sample_lines_and_extract}
\title{Sample lines and extract}
\usage{
sample_lines_and_extract(
  lines_sf,
  rast,
  d = 150,
  path_id_col = NULL,
  drop_na = TRUE
)
}
\arguments{
\item{lines_sf}{line vector to extract along}

\item{rast}{raster to extract from}

\item{d}{distance to sample line along}

\item{path_id_col}{column name for path ID}

\item{drop_na}{should NA values be dropped}
}
\value{
table with extractex values
}
\description{
Sample lines and extract
}
\examples{
# sample_lines_and_extract(lc_paths_sf, input_rast, d = 150, path_id_col = 'path_id')
}
