% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_utils.R
\name{snap_pts_to_rast}
\alias{snap_pts_to_rast}
\title{Snap points to raster}
\usage{
snap_pts_to_rast(pts, r)
}
\arguments{
\item{pts}{vector points of watershed pour points}

\item{r}{raster of cost surface}
}
\value{
vector points moved onto valid pixels of raster
}
\description{
Snap points to raster
}
\examples{
#   end_pts_moved <- snap_pts_to_rast(pts = end_points_25km, r = my_cost_surface_reachable)
}
