% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_test.R
\name{test_mannkendall}
\alias{test_mannkendall}
\title{test for directional gradient}
\usage{
test_mannkendall(df, yvar, mk_alpha = 0.05, mk_tau_threshold = 0.5)
}
\arguments{
\item{df}{data frame representing least cost path}

\item{yvar}{column name for y variable}

\item{mk_alpha}{significance threshold for MK test}

\item{mk_tau_threshold}{threshold for tau}
}
\value{
results of MK test as list
}
\description{
test for directional gradient
}
\examples{
# test_mannkendall(lc_path, yvar = 'TSM')
}
