% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateHCR_v9.R
\name{run_hcr}
\alias{run_hcr}
\title{Calculate Hydrologic Connectivity to Rivers}
\usage{
run_hcr(
  input_tif,
  sea_point,
  land_points,
  land_points_id_col = "pourpoint_id",
  output_dir,
  dist_m = 25000,
  r_min = 0,
  r_max = 200,
  mk_alpha = 0.05,
  mk_tau_threshold = 0.5,
  EXP_SCALE_PAR = 10,
  save_outputs = TRUE
)
}
\arguments{
\item{input_tif}{filepath for raster map}

\item{sea_point}{filepath for vector point of interest in coast}

\item{land_points}{filepath for vector points of river mouths}

\item{land_points_id_col}{column name for IDs of river mouths}

\item{output_dir}{folder name to save outputs}

\item{dist_m}{max distance for river mouths from sea point}

\item{r_min}{min value for standardizing raster map parameter}

\item{r_max}{max value for standardizing raster map parameter}

\item{mk_alpha}{significance level for Mann Kendall test}

\item{mk_tau_threshold}{tau threshold for Mann Kendall test}

\item{EXP_SCALE_PAR}{exponential decay parameter for cost surface}

\item{save_outputs}{save outputs or not}
}
\value{
list of results
}
\description{
Calculate Hydrologic Connectivity to Rivers
}
\examples{
# run_hcr(input_tif = 'output/01_tifs/KlamathRiver_25km/2017-01-12_S3A_tsm.tif',
 #      sea_point = 'data/geo/targetpoints/KlamathRiver_25km-5km-pt10.geojson',
 #      land_points = 'data/geo/pourpoints/KlamathRiver_25km_adj.geojson',
 #      land_points_id_col = 'pourpoint_id',
 #      output_dir = 'HCRout/klamath_10')
}
